Fix Start-RdpSession
RDP sessions have been failing since PowerShell 7 because it no longer
uses CredentialManager to add to the windows vault. After fiddling with
700 different interations of `cmdkey` with `-replace`, I found an SO
post linking back to a Microsoft support ticket that suggests using
`--%` to have a shell app stop parsing all text and treat them only as
arguments. This works a charm and allows `cmdkey` to be used with
`mstsc` as they were always meant to be used.

Related
 - https://stackoverflow.com/a/21865840/3183588