Fix WinMinimize, Activate with AHK
There was some difficulty using the UI mechanism for resolvig this issue
when switching between different users. I was able to fix it by
refereincing the exe version, but this requires that AHK run as admin.

This isn't ideal, but I currently don't start any other programs via
this method.